Add files via upload
The home page features popular businesses in UNCG and has a search bar in the center of the web page. The home page contains a interactive navigation bar that leads to each page smoothly. Shopping cart page has a table set to organize and improve the readability of the user's order.